feat: Lifting user input state to app for seamless transfer to the results component
