+++ src/App.jsx
@@ -1,11 +1,31 @@
+import { useState } from "react";
+
 import Header from "./components/Header/Header.jsx";
 import UserInput from "./components/UserInput/UserInput.jsx";
+import Results from "./components/Results/Results.jsx";
 
 function App() {
+  const [userInput, setUserInput] = useState({
+    initialInvestment: 10000,
+    annualInvestment: 1200,
+    expectedReturn: 6,
+    duration: 10,
+  });
+
+  function handleInputChange(inputIdentifier, newValue) {
+    setUserInput((prevUserInput) => {
+      return {
+        ...prevUserInput,
+        [inputIdentifier]: newValue,
+      };
+    });
+  }
+
   return (
     <>
       <Header />
-      <UserInput />
+      <UserInput userInput={userInput} onChange={handleInputChange} />
+      <Results input={userInput} />
     </>
   );
 }
