+++ src/components/Results/Results.jsx
@@ -1,0 +1,8 @@
+import "./Results.css";
+
+function Results({ input }) {
+  console.log(input);
+  return <p>Results...</p>;
+}
+
+export default Results;
