+++ src/index.css
@@ -14,7 +14,7 @@ body {
   min-height: 100vh;
 }
 
-#result {
+/* #result {
   max-width: 50rem;
   margin: 2rem auto;
   padding: 1rem;
@@ -32,7 +32,7 @@ body {
   font-family: "Roboto Condensed", sans-serif;
   font-size: 0.85rem;
   color: #c2e9e0;
-}
+} */
 
 .center {
   text-align: center;
